+++ Grupo4/src/main/java/com/Projeto/Integrador/Grupo4/Repository/BookRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.Projeto.Integrador.Grupo4.Model.BookModel;
 
-public interface BookRepository extends JpaRepository <BookModel, Integer>{
+public interface BookRepository extends JpaRepository <BookModel, Long>{
 
 	@Query("SELECT obj FROM BookModel obj WHERE UPPER(obj.title) LIKE CONCAT('%',UPPER(:title),'%')")
 	BookModel findByDescriptionTitle(@Param("title") String title);
