+++ Grupo4/src/main/java/com/Projeto/Integrador/Grupo4/Repository/CategoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.Projeto.Integrador.Grupo4.Model.CategoryModel;
 
-public interface CategoryRepository extends JpaRepository<CategoryModel,Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryModel,Long> {
 
 	@Query("SELECT obj FROM CategoryModel obj WHERE UPPER(obj.genre) LIKE CONCAT('%',UPPER(:genre),'%')")
 	CategoryModel findByDescriptionGenre(@Param("genre") String genre);
